add progress storer for wasm
