+++ core/sys/fs.go
@@ -31,6 +31,15 @@ type FS interface {
 
 	//MkdirAll creates a directory named path
 	MkdirAll(path string, perm os.FileMode) error
+
+	// LoadProgress load progress
+	LoadProgress(progressID string) ([]byte, error)
+
+	// SaveProgress save progress
+	SaveProgress(progressID string, data []byte, perm fs.FileMode) error
+
+	// RemoveProgress remove progress
+	RemoveProgress(progressID string) error
 }
 
 type File interface {
