+++ zboxcore/sdk/chunked_upload_progress_storer.go
@@ -100,7 +100,7 @@ func saveProgress(ctx context.Context, fs *fsChunkedUploadProgressStorer) {
 func (fs *fsChunkedUploadProgressStorer) Load(progressID string) *UploadProgress {
 
 	progress := UploadProgress{}
-	buf, err := sys.Files.ReadFile(progressID)
+	buf, err := sys.Files.LoadProgress(progressID)
 
 	if errors.Is(err, os.ErrNotExist) {
 		return nil
@@ -141,7 +141,7 @@ func (fs *fsChunkedUploadProgressStorer) Save(up UploadProgress) {
 		logger.Logger.Error("[progress] save ", fs.up, err)
 		return
 	}
-	err = sys.Files.WriteFile(fs.up.ID, buf, 0666)
+	err = sys.Files.SaveProgress(fs.up.ID, buf, 0666)
 	if err != nil {
 		logger.Logger.Error("[progress] save ", fs.up, err)
 		return
@@ -161,7 +161,7 @@ func (fs *fsChunkedUploadProgressStorer) Remove(progressID string) error {
 	fs.Lock()
 	defer fs.Unlock()
 	fs.isRemoved = true
-	err := sys.Files.Remove(progressID)
+	err := sys.Files.RemoveProgress(progressID)
 	if err != nil {
 		if !errors.Is(err, os.ErrNotExist) {
 			return nil
